+++ src/Layers/xrRender/dxRenderDeviceRender.h
@@ -51,7 +51,7 @@ public:
 	virtual void	DeferredLoad(BOOL E);
 	virtual void	ResourcesDeferredUpload();
 	virtual void	ResourcesDeferredUnload();
-	virtual void	ResourcesGetMemoryUsage(u32& m_base, u32& c_base, u32& m_lmaps, u32& c_lmaps);
+	virtual void	ResourcesGetMemoryUsage(size_t& m_base, u32& c_base, size_t& m_lmaps, u32& c_lmaps);
 	virtual void	ResourcesDestroyNecessaryTextures();
 	virtual void	ResourcesStoreNecessaryTextures();
 	virtual void	ResourcesDumpMemoryUsage();
